+++ examples/commonjs/build.js
@@ -36,8 +36,6 @@ module.exports = require('./lib/index')
 
 var uid = 0
 
-var processors = []
-
 var hasKeyframes = /@keyframes/
 
 var toString = Object.prototype.toString
@@ -74,37 +72,13 @@ module.exports = Rule
 Rule.NAMESPACE_PREFIX = 'jss'
 
 /**
- * Add a preprocessor.
+ * Add child rule. Required for plugins like "nested".
+ * Stylesheet will render them as a separate rule.
  *
- * @param {Function} fn
- * @return {Array}
+ * @param {String} selector
+ * @param {Object} style
+ * @return {Rule}
  * @api public
- */
-Rule.addPreprocessor = function (fn) {
-    processors.push(fn)
-    return processors
-}
-
-/**
- * Execute all registered preprocessors.
- *
- * @return {Rule}
- * @api private
- */
-Rule.prototype.runPreprocessors = function () {
-    for (var i = 0; i < processors.length; i++) {
-        processors[i](this)
-    }
-
-    return this
-}
-
-/**
- * Add child rule. Used by "nested" preprocessor.
- * Stylesheet will render them separately.
- *
- * @return {Rule}
- * @api private
  */
 Rule.prototype.addChild = function (selector, style) {
     if (!this.children) this.children = {}
@@ -157,6 +131,7 @@ Rule.prototype.toString = function () {
 'use strict'
 
 var Rule = require('./Rule')
+var plugins = require('./plugins')
 
 /**
  * Stylesheet abstraction, contains rules, injects stylesheet into dom.
@@ -321,7 +296,7 @@ Stylesheet.prototype.createRules = function (key, style) {
     rules.push(rule)
     this.rules[name || rule.selector] = rule
     if (this.named) this.classes[name] = rule.className
-    rule.runPreprocessors()
+    plugins.run(rule)
 
     for (key in rule.children) {
         rules.push(this.createRules(key, rule.children[key]))
@@ -348,7 +323,7 @@ Stylesheet.prototype.createElement = function () {
     return el
 }
 
-},{"./Rule":4}],6:[function(require,module,exports){
+},{"./Rule":4,"./plugins":7}],6:[function(require,module,exports){
 /**
  * Stylesheets written in javascript.
  *
@@ -362,13 +337,6 @@ Stylesheet.prototype.createElement = function () {
 var Stylesheet = require('./Stylesheet')
 var Rule = require('./Rule')
 
-// Register default processors.
-;[
-    require('./processors/nested'),
-    require('./processors/extend'),
-    require('./processors/vendorPrefixer')
-].forEach(Rule.addPreprocessor)
-
 exports.Stylesheet = Stylesheet
 
 exports.Rule = Rule
@@ -376,6 +344,8 @@ exports.Rule = Rule
 exports.vendorPrefix = require('./vendorPrefix')
 
 exports.support = require('./support')
+
+exports.plugins = require('./plugins')
 
 /**
  * Create a stylesheet.
@@ -399,10 +369,57 @@ exports.createStylesheet = function (rules, named, attributes) {
  * @api public
  */
 exports.createRule = function (selector, style) {
-    return new Rule(selector, style).runPreprocessors()
+    var rule = new Rule(selector, style)
+    exports.plugins.run(rule)
+    return rule
 }
 
-},{"./Rule":4,"./Stylesheet":5,"./processors/extend":7,"./processors/nested":8,"./processors/vendorPrefixer":9,"./support":10,"./vendorPrefix":11}],7:[function(require,module,exports){
+/**
+ * Register plugin. Passed function will be invoked with a rule instance.
+ *
+ * @param {Function} fn
+ * @api public
+ */
+exports.use = exports.plugins.use
+
+},{"./Rule":4,"./Stylesheet":5,"./plugins":7,"./support":11,"./vendorPrefix":12}],7:[function(require,module,exports){
+'use strict'
+
+/**
+ * Registered plugins.
+ *
+ * @type {Array}
+ * @api public
+ */
+exports.registry = [
+    require('./plugins/nested'),
+    require('./plugins/extend'),
+    require('./plugins/vendorPrefixer')
+]
+
+/**
+ * Register plugin. Passed function will be invoked with a rule instance.
+ *
+ * @param {Function} fn
+ * @api public
+ */
+exports.use = function (fn) {
+    exports.registry.push(fn)
+}
+
+/**
+ * Execute all registered plugins.
+ *
+ * @param {Rule} rule
+ * @api private
+ */
+exports.run = function (rule) {
+    for (var i = 0; i < exports.registry.length; i++) {
+        exports.registry[i](rule)
+    }
+}
+
+},{"./plugins/extend":8,"./plugins/nested":9,"./plugins/vendorPrefixer":10}],8:[function(require,module,exports){
 'use strict'
 
 var toString = Object.prototype.toString
@@ -441,7 +458,7 @@ module.exports = function (rule) {
     rule.style = newStyle
 }
 
-},{}],8:[function(require,module,exports){
+},{}],9:[function(require,module,exports){
 'use strict'
 
 var regExp = /&/gi
@@ -465,7 +482,7 @@ module.exports = function (rule) {
     }
 }
 
-},{}],9:[function(require,module,exports){
+},{}],10:[function(require,module,exports){
 'use strict'
 
 var jss = require('..')
@@ -488,7 +505,7 @@ module.exports = function (rule) {
     }
 }
 
-},{"..":6}],10:[function(require,module,exports){
+},{"..":6}],11:[function(require,module,exports){
 'use strict'
 
 var vendorPrefix = require('./vendorPrefix')
@@ -547,7 +564,7 @@ exports.getProp = function (prop) {
     return cache[prop]
 }
 
-},{"./vendorPrefix":11}],11:[function(require,module,exports){
+},{"./vendorPrefix":12}],12:[function(require,module,exports){
 'use strict'
 
 var jsCssMap = {